Made weight tags controllable from admin panel
