+++ communityEmpowerment/apps.py
@@ -3,3 +3,5 @@ from django.apps import AppConfig
 
 class communityEmpowermentConfig(AppConfig):
     name = 'communityEmpowerment'
+    def ready(self):
+        import communityEmpowerment.signals
